requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg[binary]==3.2.9
python-dotenv==1.1.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0.post1
python-multipart>=0.0.9
